BUILD: log/threads: implement the logsrv lock
Commit 0a45fd2b2 ("BUG/MINOR: log: make sure writev() is not interrupted
on a file output") relies on the logsrv's lock, which doesn't exist in 1.9
and which was introduced by commit d803e475e ("MINOR: log: Enable the log
sampling and load-balancing feature."). As such the commit above breaks
the build in 1.9 with threads.

This patch is only a minimalistic implementation of the logsrv lock, which
contains only the declaration and initialization. It doesn't have a mainline
equivalent though it can be seen as a reduced backport of d803e475e.

Signed-off-by: Willy Tarreau <w@1wt.eu>