BUG/MEDIUM: fd: Always reset the polled_mask bits in fd_dodelete().
In fd_dodelete(), always reset the polled_mask bits, instead on only doing
it if we're closing the file descriptor. We call the poller clo() method
anyway, and failing to do so means that if fd_remove() is used while the
fd is polled, the poller won't attempt to poll on a fd with the same value
as the old one.
This leads to fd being stuck in the SSL code while using the async engine.

This should be backported to 2.0, 1.9 and 1.8.

(cherry picked from commit c22580c2cce86473221f58cf11db19056a5f2f7c)
[wt: polled_mask[] is only an array in 2.0]
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 15a2e6856e5ac7b5de1f6c1b6449dfda12daa138)
Signed-off-by: Willy Tarreau <w@1wt.eu>