BUG/MINOR: cache: alloc shctx after check config
When running haproxy -c, the cache parser is trying to allocate the size
of the cache. This can be a problem in an environment where the RAM is
limited.

This patch moves the cache allocation in the post_check callback which
is not executed during a -c.

This patch may be backported at least to 2.0 and 1.9. In 1.9, the callbacks
registration mechanism is not the same. So the patch will have to be adapted. No
need to backport it to 1.8, the code is probably too different.

(cherry picked from commit d1d1e229453a492a538245f6a72ba6929eca9de1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit e4876e2b03930a5e280c77f9ebd59f861080a3c5)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>