BUG/MINOR: stick-table: Never exceed (MAX_SESS_STKCTR-1) when fetching a stkctr
When a stick counter is fetched, it is important that the requested counter does
not exceed (MAX_SESS_STKCTR -1). Actually, there is no bug with a default build
because, by construction, MAX_SESS_STKCTR is defined to 3 and we know that we
never exceed the max value. scN_* sample fetches are numbered from 0 to 2. For
other sample fetches, the value is tested.

But there is a bug if MAX_SESS_STKCTR is set to a lower value. For instance
1. In this case the counters sc1_* and sc2_* may be undefined.

This patch fixes the issue #330. It must be backported as far as 1.7.

(cherry picked from commit a9fa88a1eac9bd0ad2cfb761c4b69fd500a1b056)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 33c7e12479cb9bdc2e7e3783fda78a1b2c242363)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>