BUG/MAJOR: mux-h2: fix incorrect backport of connection timeout fix
In commit c2ea47f ("BUG/MEDIUM: mux-h2: do not enforce timeout on
long connections") the changes were applied to the proper location but
while this code was moved before the task_destroy() call in 2.0, in 1.9
it is after task_free() so we can return a task that was just freed, in
case the timeout strikes and at the same time a new stream arrives.

The probability that this issue occurs is rather low but its consequences
are high and this could be the cause of issue #329.

This patch is specific to 1.9 as the issue does not exist in newer
versions. If the commit above is ever backported to 1.8, this one will
need to as well.