BUG/MINOR: cli: don't call the kw->io_release if kw->parse failed
The io_release() callback of the cli_kw is supposed to be used to clean
what an io_handler() has made. It is called once the work in the IO
handler is finished, or when the connection was aborted by the client.

This patch fixes a bug where the io_release callback was called even
when the parse() callback failed. Which means that the io_release() could
called even if the io_handler() was not called.

Should be backported in every versions that have a cli_kw->release().
(as far as 1.7)

(cherry picked from commit 90b098c921e15f912dbde42658e34780f0ba446d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit d4f20fadd9c3145de0eb5f5434f57b9fffc61062)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>