BUG/MEDIUM: stream: Be sure to support splicing at the mux level to enable it
Despite the addition of the mux layer, no change have been made on how to enable
the TCP splicing on process_stream(). We still check if transport layer on both
sides support the splicing, but we don't check the muxes support. So it is
possible to start to splice data with an unencrypted H2 connection on a side and
an H1 connection on the other. This leads to a freeze of the stream until a
client or server timeout is reached.

This patch fixed a part of the issue #356. It must be backported as far as 1.8.

(cherry picked from commit 276c1e0533e77008445d57a1953f1f516d66877d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 22b4eefa969bb741d75e78b5da457b0322213f69)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>