BUG: dns: timeout resolve not applied for valid resolutions
Documentation states that the interval between 2 DNS resolution is
driven by "timeout resolve <time>" directive.
From a code point of view, this was applied unless the latest status of
the resolution was VALID. In such case, "hold valid" was enforce.
This is a bug, because "hold" timers are not here to drive how often we
want to trigger a DNS resolution, but more how long we want to keep an
information if the status of the resolution itself as changed.
This avoid flapping and prevent shutting down an entire backend when a
DNS server is not answering.

This issue was reported by hamshiva in github issue #345.

Backport status: 1.8

(cherry picked from commit f50e1ac4442be41ed8b9b7372310d1d068b85b33)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 3697f03e859f5af5a2d759e6c3d9e2816b61871c)
Signed-off-by: Willy Tarreau <w@1wt.eu>